Move CopilotModelSelectionInfo to its own file
